fixed issue with wxFormatString ni configview.cpp when compiling against wxWidgets 2.9.1
git-svn-id: https://usbpicprog.svn.sourceforge.net/svnroot/usbpicprog@923 29735491-9d32-0410-b17f-ed4c7e910dea

+++ trunk/upp_wx/src/configview.cpp
@@ -166,7 +166,7 @@ void UppConfigViewBook::SetHexFile(HexFile* hex, const PicType& pic)
         // add a few more details about the location of current config word in PIC memory
 
         boxSz->AddStretchSpacer(1);
-        boxSz->Add(new wxStaticText(panel, wxID_ANY, wxString::Format(_("Location in memory: 0x%X"), m_pic.ConfigAddress + word.Offset)),
+        boxSz->Add(new wxStaticText(panel, wxID_ANY, wxString::Format(_("Location in memory: 0x%X"), m_pic.ConfigAddress +(int) word.Offset)),
                 0, wxLEFT|wxALIGN_CENTER, 5);
                 
         boxSz->AddStretchSpacer(1);
